WIP: refactoring to reduce duplicated code
